ROB-320 Claude/remove edit command step2 (#2202)
